improve serial port selection behavior
When no devices are connect the only ports in the list are the system
COM ports (ttyS0) under linux. Now when a mesytec device is connected the
combo box did not auto select the device but instead kept the COM port
selected.

After this change the first serial port that has a serial number is
selected if the currently selected port does not have a serial number.